Create MaterialSelect.vue
Select component to use with Vue.js and MaterializeCSS. The 'select' in MaterializeCSS converts it to an 'ul' this cause that the data binding use in Vue.js don’t work. This component is a workaround for the select, it use jQuery to add, update, and listen to changes of the select.